Create infrastructure:crypto module Move ResourceLoader and related classes to infrastructure:io
