+++ ethereum/datastructures/build.gradle
@@ -6,6 +6,7 @@ dependencies {
   implementation project(':bls')
   implementation project(':infrastructure:async')
   implementation project(':infrastructure:collections')
+  implementation project(':infrastructure:io')
   implementation project(':infrastructure:logging')
   implementation project(':pow')
   implementation project(':util')
