+++ infrastructure/crypto/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+  api 'org.bouncycastle:bcprov-jdk15on'
+
+}
