+++ settings.gradle
@@ -16,6 +16,7 @@ include 'fuzz'
 include 'infrastructure'
 include 'infrastructure:async'
 include 'infrastructure:collections'
+include 'infrastructure:crypto'
 include 'infrastructure:events'
 include 'infrastructure:io'
 include 'infrastructure:metrics'
