+++ util/build.gradle
@@ -3,12 +3,13 @@ plugins {
 }
 
 dependencies {
-  api 'org.bouncycastle:bcprov-jdk15on'
+  api project(':infrastructure:crypto')
   api project(':infrastructure:metrics')
   api project(':infrastructure:events')
 
   implementation project(':bls')
   implementation project(':infrastructure:async')
+  implementation project(':infrastructure:io')
   implementation project(':infrastructure:logging')
   implementation project(':infrastructure:subscribers')
   implementation project(':ssz')
