+++ util/src/main/java/tech/pegasys/teku/util/config/Constants.java
@@ -19,9 +19,9 @@ import java.io.InputStream;
 import java.util.concurrent.TimeUnit;
 import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
+import tech.pegasys.teku.infrastructure.io.resource.ResourceLoader;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 import tech.pegasys.teku.ssz.SSZTypes.Bytes4;
-import tech.pegasys.teku.util.resource.ResourceLoader;
 
 public class Constants {
